fix updateById bug & more cases
